+++ src/info.py
@@ -1,4 +1,4 @@
 class Info:
     AppName = 'S-Parameter Viewer'
     AppVersionStr = "0.27b1" # see https://peps.python.org/pep-0440/
-    AppDateStr = "2025-03-01"
+    AppDateStr = "2025-03-02"