+++ CHANGELOG.md
@@ -7,6 +7,7 @@ S-Parameter Viewer Changelog
 
 - new: function `vswr()` to calculate VSWR
 - new: new sample `dipole.s1p`
+- new: histogram in optimum RL calculator
 - bugfix: no longer showing losslessness ij metric for 1-ports
 
 0.25b2 (2025-02-23)